requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
anyio==4.6.2
    # via httpx
authlib==1.3.2
    # via -r requirements.in
blinker==1.8.2
    # via flask
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.7
    # via flask
cryptography==43.0.1
    # via authlib
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-wtf
flask-login==0.6.3
    # via -r requirements.in
flask-wtf==1.2.1
    # via -r requirements.in
flickr-photos-api==2.11.2
    # via -r requirements.in
flickr-url-parser==1.10.0
    # via flickr-photos-api
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via
    #   flickr-photos-api
    #   flickr-url-parser
hyperlink==21.0.0
    # via flickr-url-parser
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jinja2==3.1.4
    # via flask
keyring==25.4.1
    # via -r requirements.in
markupsafe==3.0.1
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
more-itertools==10.5.0
    # via
    #   jaraco-classes
    #   jaraco-functools
pycparser==2.22
    # via cffi
silver-nitrate==1.1.3
    # via flickr-photos-api
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==9.0.0
    # via flickr-photos-api
werkzeug==3.0.4
    # via
    #   flask
    #   flask-login
wtforms==3.1.2
    # via flask-wtf
